+++ implementation/src/main/kotlin/io/github/tomplum/aoc/bridge/Operation.kt
@@ -3,7 +3,7 @@ package io.github.tomplum.aoc.bridge
 enum class Operation {
     ADD, MULTIPLY;
 
-    fun apply(first: Int, second: Int) = when (this) {
+    fun apply(first: Long, second: Long) = when (this) {
         ADD -> first + second
         MULTIPLY -> first * second
     }
