+++ implementation/src/main/kotlin/io/github/tomplum/aoc/bridge/RopeBridgeSchematics.kt
@@ -4,7 +4,7 @@ class RopeBridgeSchematics(input: List<String>) {
 
     private val equations = input.map { line -> CalibrationEquation(line)}
 
-    fun getTotalCalibrationResult(): Int = equations
+    fun getTotalCalibrationResult(): Long = equations
         .filter { equation -> equation.hasValidOperatorConfig() }
         .sumOf { equation -> equation.testValue
     }
